requirements.txt
Django==2.1.3
pytest-django==3.4.4
django-environ==0.4.5
psycopg2==2.7.6.1
gunicorn==19.9.0
whitenoise==4.1.2
djangorestframework==3.9.0
